requirements.txt
streamlit
python-digitalocean
